Fix long

+++ src/EmployeVente.java
@@ -1,6 +1,6 @@
 public class EmployeVente extends Employe{
     private double chiffreAffaire;
-    private double vat;
+    private long vat;
 
     public EmployeVente(String prenom, String nom, int age, double chiffreAffaire){
         super(prenom, nom, age);
@@ -22,11 +22,11 @@ public class EmployeVente extends Employe{
         this.chiffreAffaire = chiffreAffaire;
     }
 
-    public double getVat() {
+    public long getVat() {
         return vat;
     }
 
-    public void setVat(double vat) {
+    public void setVat(long vat) {
         this.vat = vat;
     }
 }
